+++ app/src/androidTest/kotlin/org/wordpress/aztec/demo/tests/MixedTextFormattingTests.kt
@@ -68,8 +68,10 @@ class MixedTextFormattingTests : BaseTest() {
 
     @Test
     fun testRetainParagraphFormatting() {
+        // Paragraph are only retained if they have at least one attribute
+        // https://github.com/wordpress-mobile/AztecEditor-Android/pull/483#discussion_r151377857
         val text = "some text"
-        val html = "<p>$text</p>"
+        val html = "<p a=\"ok\">$text</p>"
 
         EditorPage()
                 .toggleHtml()